requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
appdirs==1.4.3            # via setuptools
beautifulsoup4==4.5.3     # via wagtail
boto==2.39.0
dj-database-url==0.4.0
django-debug-toolbar==1.4
django-modelcluster==3.0.1  # via wagtail
django-modeltranslation==0.12  # via wagtail-modeltranslation
django-nose==1.4.3
django-storages==1.4
django-taggit==0.22.0     # via wagtail
django-treebeard==4.1.0   # via wagtail
django-wagtail-feeds==0.0.6
Django==1.9.12            # via django-debug-toolbar, django-treebeard, django-wagtail-feeds, wagtail
djangorestframework==3.6.2  # via wagtail
elasticsearch==2.3.0
gunicorn==19.4.5
html5lib==0.999999999     # via wagtail
nose==1.3.7               # via django-nose
olefile==0.44             # via pillow
packaging==16.8           # via setuptools
Pillow==4.0.0             # via wagtail
psycopg2==2.6.1
pyparsing==2.2.0          # via packaging
pytz==2016.10             # via django-modelcluster
requests==2.13.0          # via wagtail
six==1.10.0               # via html5lib, packaging, setuptools
sqlparse==0.2.3           # via django-debug-toolbar
Unidecode==0.4.20         # via wagtail
urllib3==1.20             # via elasticsearch
wagtail-modeltranslation==0.6.0rc1
wagtail==1.9
webencodings==0.5         # via html5lib
whitenoise==2.0.6
Willow==0.4               # via wagtail

# The following packages are considered to be unsafe in a requirements file:
# setuptools                # via html5lib
